refactor(markdown): Move `astro:assets`-specific code out of the main Vite plugin
